beta 1.0.2
หน้าผู้ใช้
- ปรับปรุงหน้าการคำนวณระยะเวลาในการลดน้ำหนัก ให้อ้างอิงจากน้ำหนักปัจจุบัน
หน้าแคลลอรี่
- แก้ไข progessbar ให้สามารถคำนวณหลอดได้อย่างถูกต้อง
- เพิ่ม เมื่อทำการกำหนดเป้าหมายน้ำหนัก ผู้ใช้จะไม่สามารถตั้งเป้าหมายนน้ำหนักให้ต่ำกว่า BMI
- สำหรับแคลลอรี่ที่ไม่มีอยู่ในระบบ จะไม่ขึ้นปุ่มให้แก้ไขรายการอาหาร

+++ app/src/main/java/com/example/james/foodandrunning/CalorieFragment.kt
@@ -77,7 +77,7 @@ class CalorieFragment : Fragment() {
             sumCalCalories(sex, weight, routine, age)
         }
 
-        if (calories == 0 && sex != 0 && weight != 0 && routine != 0 && age != 0) {
+        if (calories == 0 && sex != 0 && weight != 0f && routine != 0 && age != 0) {
            sumCalCalories(sex, weight, routine, age)
         }
 
@@ -379,7 +379,7 @@ class CalorieFragment : Fragment() {
 
     private fun sumCalCalories(
         sex: Int,
-        weight: Int,
+        weight: Float,
         routinein: Int,
         age: Int
     ) {
@@ -446,7 +446,7 @@ class CalorieFragment : Fragment() {
         FirebaseFirestore.getInstance().collection("WEIGHT_TABLE").document(weightpath).get().addOnSuccessListener {
             if (it != null) {
                 val dataHash = it.data as HashMap<String, Any>
-                val weight = dataHash["weight_value"].toString().toInt()
+                val weight = dataHash["weight_value"].toString().toFloat()
                 AppPreferences(this.context!!).setPreferenceWeight(weight)
             }
         }
